refactor:Implement sorting and comparison functions
